refactor(conformance): split error_handling checker into package + enable pyright
Split the 1147-line conformance/suite/checks/error_handling.py monolith into
a proper package (error_handling/) with one module per category:
- _constants.py   — shared frozensets and compiled regexes
- _directives.py  — ignore-directive parsing
- _collect.py     — AST import-collection helpers
- _helpers.py     — shared AST utilities and is_builtin_raise/is_broad_suppress
- silent_swallow.py / untyped_raise.py / exception_chaining.py / security.py
  — one mixin per E-series rule family
- _checker.py     — Checker class composing the four mixins + ast.NodeVisitor
- __init__.py     — public API: SERIES, discover, scan_path, scan_text, main

Public symbols (BUILTIN_RAISES, LEAF_CLASSES, LEGACY_ATLAN_ERRORS, scan_text,
parse_ignore_directive, is_broad_suppress, is_builtin_raise, main) are
re-exported unchanged so all 213 tests pass without modification.

Also remove the conformance/** exclusion from pyright (pyproject.toml) and add
extraPaths = ["conformance"] so the suite is fully type-checked. Fix the
resulting issues:

- sarif.py: replace per-field alias= with alias_generator=to_camel across all
  models (eliminates pyright's alias-overload requirement for all constructor
  args); inject $schema via model_serializer on SarifReport (the $ prefix
  can't be derived from to_camel)
- builder.py: update all Pydantic constructor calls to use Python snake_case
  names; fix add_pass Location/Region construction
- catalog.py: already used Python names — now accepted by pyright after the
  model fix
- test_disposition.py: update constructor calls to Python names; add explicit
  start_column=1 for Region (alias-generator overload)
- test_error_handling.py: add None guards for Optional[PhysicalLocation] chain

Upgrade pyright pre-commit hook from v1.1.393 → v1.1.410.
