+++ conformance/suite/checks/error_handling/silent_swallow.py
@@ -1,0 +1,313 @@
+"""E001–E011 and E014 — silent exception swallowing rules."""
+
+from __future__ import annotations
+
+import ast
+
+from ._constants import (
+    _BROAD_EXCEPT_TYPES,
+    _OPTIONAL_IMPORT_TYPES,
+)
+from ._helpers import (
+    _any_logging_in,
+    _body_is_only_loop_control_no_logging,
+    _body_is_only_pass,
+    _filter_body_wrapped,
+    _find_filter_method,
+    _get_name,
+    _has_exc_info,
+    _inherits_logging_filter,
+    _is_gather_call,
+    _iter_shallow,
+)
+
+
+class SilentSwallowMixin:
+    """Rule methods for E001–E011 and E014 (silent-swallow category)."""
+
+    # ── E001 / E002 / E006 ───────────────────────────────────────────────────
+
+    def _check_p001_p002_p006(self, node: ast.ExceptHandler) -> None:
+        is_bare = node.type is None
+        is_pass_only = _body_is_only_pass(node.body)
+
+        if is_bare:
+            if is_pass_only:
+                self._add(
+                    "E001",
+                    node,
+                    "Bare 'except: pass' silently discards every exception including "
+                    "SystemExit and KeyboardInterrupt — the hardest class of bugs to debug. "
+                    "Replace with a typed catch that at minimum logs at DEBUG.",
+                )
+            else:
+                self._add(
+                    "E006",
+                    node,
+                    "Bare 'except:' (no type) catches SystemExit and KeyboardInterrupt. "
+                    "Use 'except Exception:' at minimum.",
+                )
+        elif is_pass_only:
+            exc_type = _get_name(node.type) or "Exception"
+            self._add(
+                "E002",
+                node,
+                f"'except {exc_type}: pass' silently discards the exception with no trace. "
+                "Acceptable only for genuinely trivial best-effort operations — "
+                "add a comment and log at DEBUG or use the suppression directive.",
+            )
+
+    # ── E003 ─────────────────────────────────────────────────────────────────
+
+    def _check_p003(self, node: ast.Call) -> None:
+        name = _get_name(node.func)
+        if name != "suppress":
+            return
+        for arg in node.args:
+            arg_name = _get_name(arg)
+            if arg_name in _BROAD_EXCEPT_TYPES:
+                self._add(
+                    "E003",
+                    node,
+                    f"contextlib.suppress({arg_name}) — scope is too broad; "
+                    f"suppresses every exception. Use a specific exception type "
+                    f"(e.g. suppress(FileNotFoundError)).",
+                )
+                return
+
+    # ── E004 ─────────────────────────────────────────────────────────────────
+
+    def _check_p004(self, node: ast.ExceptHandler) -> None:
+        if node.type is None:
+            return  # bare except handled by E006
+        if isinstance(node.type, ast.Tuple):
+            broad = {_get_name(e) for e in node.type.elts} & _BROAD_EXCEPT_TYPES
+        else:
+            name = _get_name(node.type)
+            broad = {name} & _BROAD_EXCEPT_TYPES if name else set()
+        if not broad:
+            return
+        exc_type = next(iter(broad))
+        # Pass if body has logger.exception() or any log call with exc_info=True
+        for n in _iter_shallow(node):
+            if not isinstance(n, ast.Expr):
+                continue
+            call = n.value
+            if isinstance(call, ast.Await):
+                call = call.value
+            if not isinstance(call, ast.Call):
+                continue
+            func = call.func
+            if not isinstance(func, ast.Attribute):
+                continue
+            if func.attr == "exception":
+                return
+            if func.attr in ("warning", "error", "critical") and _has_exc_info(call):
+                return
+        self._add(
+            "E004",
+            node,
+            f"'except {exc_type}' catches everything. Acceptable only at top-level handlers "
+            f"(worker loops, HTTP handlers) when logged with exc_info=True. "
+            f"Narrow the exception type or add exc_info=True logging.",
+        )
+
+    # ── E005 ─────────────────────────────────────────────────────────────────
+
+    def _check_p005(self, node: ast.ExceptHandler) -> None:
+        for n in _iter_shallow(node):
+            if n is node:
+                continue
+            if not isinstance(n, ast.Expr):
+                continue
+            call = n.value
+            if isinstance(call, ast.Await):
+                call = call.value
+            if not isinstance(call, ast.Call):
+                continue
+            func = call.func
+            if not isinstance(func, ast.Attribute):
+                continue
+            if func.attr == "exception":
+                continue  # logger.exception() implies exc_info — skip
+            if func.attr in ("warning", "error", "critical") and not _has_exc_info(
+                call
+            ):
+                self._add(
+                    "E005",
+                    n,
+                    f"logger.{func.attr}() inside except block is missing exc_info=True — "
+                    f"the stack trace is silently discarded. Add exc_info=True.",
+                )
+
+    # ── E007 ─────────────────────────────────────────────────────────────────
+
+    def _check_p007(self, node: ast.ExceptHandler) -> None:
+        for i, stmt in enumerate(node.body):
+            if not isinstance(stmt, ast.Return) or stmt.value is None:
+                continue
+            if _any_logging_in(node.body[:i]):
+                continue
+            self._add(
+                "E007",
+                stmt,
+                "except block returns a value without logging — the error is hidden. "
+                "Log before returning or raise a domain-specific exception.",
+            )
+
+    # ── E008 ─────────────────────────────────────────────────────────────────
+
+    def _check_p008(self, node: ast.ExceptHandler) -> None:
+        if node.type is None:
+            return
+        exc_type = _get_name(node.type)
+        if exc_type not in _OPTIONAL_IMPORT_TYPES:
+            return
+        if _any_logging_in(node.body):
+            return
+        self._add(
+            "E008",
+            node,
+            f"'except {exc_type}' with no logging — import failures are silently hidden. "
+            f"Log at DEBUG if the module is preferred but not required, or add a comment "
+            f"explaining the optional dependency and use the suppression directive.",
+        )
+
+    # ── E009 ─────────────────────────────────────────────────────────────────
+
+    def _check_p009(self, node: ast.ExceptHandler) -> None:
+        real_stmts = [
+            s
+            for s in node.body
+            if not (
+                isinstance(s, ast.Expr)
+                and isinstance(s.value, ast.Constant)
+                and isinstance(s.value.value, str)
+            )
+        ]
+        if not real_stmts:
+            return
+        if not all(
+            isinstance(s, (ast.Assign, ast.AugAssign, ast.AnnAssign))
+            for s in real_stmts
+        ):
+            return
+        if _any_logging_in(node.body):
+            return
+        self._add(
+            "E009",
+            node,
+            "except block only assigns a variable — the exception is silently hidden. "
+            "Add logger.warning(..., exc_info=True) before the assignment.",
+        )
+
+    # ── E010 ─────────────────────────────────────────────────────────────────
+
+    def _check_p010_in_function(
+        self, func_node: ast.FunctionDef | ast.AsyncFunctionDef
+    ) -> None:
+        """Flag asyncio.gather(return_exceptions=True) whose results are not inspected."""
+        bare_gathers: list[ast.AST] = []
+        gather_vars: dict[str, ast.AST] = {}
+
+        for node in _iter_shallow(func_node):
+            # Bare expression: gather result discarded entirely
+            if isinstance(node, ast.Expr):
+                val = node.value
+                if isinstance(val, ast.Await):
+                    val = val.value
+                if isinstance(val, ast.Call) and _is_gather_call(val):
+                    for kw in val.keywords:
+                        if kw.arg == "return_exceptions":
+                            v = kw.value
+                            if isinstance(v, ast.Constant) and v.value is True:
+                                bare_gathers.append(node)
+                continue
+            # Assigned: results = await asyncio.gather(..., return_exceptions=True)
+            if isinstance(node, ast.Assign):
+                val = node.value
+                if isinstance(val, ast.Await):
+                    val = val.value
+                if isinstance(val, ast.Call) and _is_gather_call(val):
+                    has_re = any(
+                        kw.arg == "return_exceptions"
+                        and isinstance(kw.value, ast.Constant)
+                        and kw.value.value is True
+                        for kw in val.keywords
+                    )
+                    if has_re:
+                        # Only track single Name targets: chained assignments
+                        # (a = b = ...) would emit one finding per target, and
+                        # attribute targets (self.x = ...) produce false positives
+                        # because the inspection-side check only matches ast.Name.
+                        if len(node.targets) == 1 and isinstance(
+                            node.targets[0], ast.Name
+                        ):
+                            gather_vars[node.targets[0].id] = node
+
+        for node in bare_gathers:
+            self._add(
+                "E010",
+                node,
+                "asyncio.gather(return_exceptions=True) result is discarded — exception "
+                "instances in the result list are silently ignored. "
+                "Inspect results: 'for r in results: if isinstance(r, Exception): ...'",
+            )
+
+        for var_name, assign_node in gather_vars.items():
+            inspected = False
+            for node in _iter_shallow(func_node):
+                # isinstance(var, ...) direct check
+                if isinstance(node, ast.Call):
+                    func = node.func
+                    if isinstance(func, ast.Name) and func.id == "isinstance":
+                        if node.args and isinstance(node.args[0], ast.Name):
+                            if node.args[0].id == var_name:
+                                inspected = True
+                                break
+                # for r in var: ... — iteration counts as inspection
+                if isinstance(node, ast.For):
+                    if isinstance(node.iter, ast.Name) and node.iter.id == var_name:
+                        inspected = True
+                        break
+            if not inspected:
+                self._add(
+                    "E010",
+                    assign_node,
+                    f"asyncio.gather(return_exceptions=True) result '{var_name}' is not "
+                    f"inspected for exception instances — errors vanish silently. "
+                    f"Check each result: 'for r in {var_name}: if isinstance(r, Exception): ...'",
+                )
+
+    # ── E011 ─────────────────────────────────────────────────────────────────
+
+    def _check_p011(self, cls: ast.ClassDef) -> None:
+        if not _inherits_logging_filter(cls):
+            return
+        method = _find_filter_method(cls)
+        if method is None:
+            return
+        if not _filter_body_wrapped(method):
+            self._add(
+                "E011",
+                method,
+                f"logging.Filter.filter() body in '{cls.name}' is not fully wrapped in "
+                f"try/except — an unguarded exception crashes the logging caller "
+                f"(Logger.handle() has no try/except around filters). "
+                f"Wrap the entire body and return a safe fallback (True = pass-through).",
+            )
+
+    # ── E014 ─────────────────────────────────────────────────────────────────
+
+    def _check_p014(self, node: ast.ExceptHandler) -> None:
+        if not self._loop_stack:
+            return
+        if not _body_is_only_loop_control_no_logging(node.body):
+            return
+        exc_type = _get_name(node.type) if node.type else "(bare)"
+        self._add(
+            "E014",
+            node,
+            f"except {exc_type}: [continue/break/pass] inside a loop — exception is "
+            f"silently swallowed. Log at DEBUG before the loop control statement.",
+        )
