+++ conformance/suite/schema/sarif.py
@@ -13,6 +13,11 @@ Design notes
   level by the typed ``AtlanRuleProperties`` / ``AtlanResultProperties``
   helpers in ``extensions.py``.
 * Enum values are the exact lowercase strings SARIF 2.1.0 requires.
+* ``model_config`` uses ``alias_generator=to_camel`` so all snake_case fields
+  serialize to camelCase SARIF wire names automatically.  ``populate_by_name``
+  lets ``model_validate`` accept either name.  Constructors use Python names.
+* ``SarifReport.schema_`` uses an explicit ``serialization_alias="$schema"``
+  because the ``$`` prefix cannot be derived from ``to_camel``.
 
 SARIF reference: https://docs.oasis-open.org/sarif/sarif/v2.1.0/
 """
@@ -21,7 +26,12 @@ from __future__ import annotations
 
 from typing import Any
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, ConfigDict, Field, model_serializer
+from pydantic.alias_generators import to_camel
+
+_SARIF_SCHEMA_URL = "https://raw.githubusercontent.com/oasis-tcs/sarif-spec/master/Schemata/sarif-schema-2.1.0.json"
+
+_SARIF_COMMON = ConfigDict(populate_by_name=True, alias_generator=to_camel)
 
 # ---------------------------------------------------------------------------
 # Location
@@ -31,14 +41,14 @@ from pydantic import BaseModel, Field
 class Region(BaseModel):
     """A region within an artifact (file)."""
 
-    start_line: int = Field(..., alias="startLine", ge=1)
-    start_column: int = Field(1, alias="startColumn", ge=1)
-    end_line: int | None = Field(None, alias="endLine")
-    end_column: int | None = Field(None, alias="endColumn")
+    start_line: int = Field(..., ge=1)
+    start_column: int = Field(1, ge=1)
+    end_line: int | None = None
+    end_column: int | None = None
     snippet: dict[str, str] | None = None
     """Optional ``{text: "..."}`` snippet for context."""
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 class ArtifactLocation(BaseModel):
@@ -47,29 +57,29 @@ class ArtifactLocation(BaseModel):
     uri: str
     """Relative or absolute URI.  Use repo-root-relative paths for portability."""
 
-    uri_base_id: str | None = Field(None, alias="uriBaseId")
+    uri_base_id: str | None = None
     """Symbolic root, e.g. ``%SRCROOT%``."""
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 class PhysicalLocation(BaseModel):
     """A physical location within a source file."""
 
-    artifact_location: ArtifactLocation = Field(..., alias="artifactLocation")
+    artifact_location: ArtifactLocation
     region: Region | None = None
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 class Location(BaseModel):
     """A location associated with a result."""
 
-    physical_location: PhysicalLocation | None = Field(None, alias="physicalLocation")
+    physical_location: PhysicalLocation | None = None
     message: dict[str, str] | None = None
     """Optional ``{text: "..."}`` message attached to the location."""
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 # ---------------------------------------------------------------------------
@@ -110,20 +120,20 @@ class Suppression(BaseModel):
 class Replacement(BaseModel):
     """A single text replacement within a file."""
 
-    deleted_region: Region = Field(..., alias="deletedRegion")
-    inserted_content: dict[str, str] | None = Field(None, alias="insertedContent")
+    deleted_region: Region
+    inserted_content: dict[str, str] | None = None
     """``{text: "replacement text"}``."""
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 class ArtifactChange(BaseModel):
     """Proposed changes to a single artifact."""
 
-    artifact_location: ArtifactLocation = Field(..., alias="artifactLocation")
+    artifact_location: ArtifactLocation
     replacements: list[Replacement] = []
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 class Fix(BaseModel):
@@ -132,9 +142,9 @@ class Fix(BaseModel):
     description: dict[str, str] | None = None
     """``{text: "..."}`` description of what the fix does."""
 
-    artifact_changes: list[ArtifactChange] = Field([], alias="artifactChanges")
+    artifact_changes: list[ArtifactChange] = []
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 # ---------------------------------------------------------------------------
@@ -156,10 +166,10 @@ class Result(BaseModel):
     * ``kind="fail"``, ``suppressions`` non-empty → SUPPRESSED (own category)
     """
 
-    rule_id: str = Field(..., alias="ruleId")
+    rule_id: str
     """Must match a ``reportingDescriptor.id`` in ``tool.driver.rules``."""
 
-    rule_index: int | None = Field(None, alias="ruleIndex")
+    rule_index: int | None = None
     """Zero-based index into ``tool.driver.rules`` for fast lookup."""
 
     kind: str = "fail"
@@ -176,9 +186,7 @@ class Result(BaseModel):
     locations: list[Location] = []
     """File+line locations; typically one entry per result."""
 
-    partial_fingerprints: dict[str, str] = Field(
-        default_factory=dict, alias="partialFingerprints"
-    )
+    partial_fingerprints: dict[str, str] = Field(default_factory=dict)
     """Stable identity hashes for deduplication and oscillation detection.
     Key ``"atlanConformance/v1"`` = ``hash(ruleId + normalised_uri + startLine)``."""
 
@@ -191,7 +199,7 @@ class Result(BaseModel):
     properties: dict[str, Any] = Field(default_factory=dict)
     """``atlan/*`` extension properties; see ``AtlanResultProperties``."""
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 # ---------------------------------------------------------------------------
@@ -207,7 +215,7 @@ class ReportingConfiguration(BaseModel):
 
     enabled: bool = True
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 class ReportingDescriptor(BaseModel):
@@ -223,24 +231,23 @@ class ReportingDescriptor(BaseModel):
     name: str | None = None
     """Short CamelCase name, e.g. ``"NoFStringInLog"``."""
 
-    short_description: dict[str, str] | None = Field(None, alias="shortDescription")
+    short_description: dict[str, str] | None = None
     """``{text: "one-liner"}``."""
 
-    full_description: dict[str, str] | None = Field(None, alias="fullDescription")
+    full_description: dict[str, str] | None = None
     """``{text/markdown: "extended description"}``."""
 
-    help_uri: str | None = Field(None, alias="helpUri")
+    help_uri: str | None = None
     """Link to rule documentation."""
 
     default_configuration: ReportingConfiguration = Field(
-        default_factory=ReportingConfiguration,
-        alias="defaultConfiguration",
+        default_factory=ReportingConfiguration
     )
 
     properties: dict[str, Any] = Field(default_factory=dict)
     """``atlan/*`` governance fields; see ``AtlanRuleProperties``."""
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 # ---------------------------------------------------------------------------
@@ -253,11 +260,11 @@ class ToolComponent(BaseModel):
 
     name: str
     version: str | None = None
-    semantic_version: str | None = Field(None, alias="semanticVersion")
-    information_uri: str | None = Field(None, alias="informationUri")
+    semantic_version: str | None = None
+    information_uri: str | None = None
     rules: list[ReportingDescriptor] = []
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 class Tool(BaseModel):
@@ -265,7 +272,7 @@ class Tool(BaseModel):
 
     driver: ToolComponent
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 # ---------------------------------------------------------------------------
@@ -276,17 +283,17 @@ class Tool(BaseModel):
 class Invocation(BaseModel):
     """Records how and when the suite was invoked."""
 
-    execution_successful: bool = Field(True, alias="executionSuccessful")
+    execution_successful: bool = True
     """``True`` if the suite ran to completion (regardless of violations found)."""
 
-    exit_code: int | None = Field(None, alias="exitCode")
+    exit_code: int | None = None
     """0 = gate passed; 1 = gate failed (≥1 FAILING result)."""
 
-    exit_code_description: str | None = Field(None, alias="exitCodeDescription")
+    exit_code_description: str | None = None
 
-    working_directory: ArtifactLocation | None = Field(None, alias="workingDirectory")
+    working_directory: ArtifactLocation | None = None
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 # ---------------------------------------------------------------------------
@@ -297,14 +304,14 @@ class Invocation(BaseModel):
 class VersionControlDetails(BaseModel):
     """Identifies the repo + commit being analysed."""
 
-    repository_uri: str = Field(..., alias="repositoryUri")
-    revision_id: str | None = Field(None, alias="revisionId")
+    repository_uri: str
+    revision_id: str | None = None
     """Commit SHA."""
 
     branch: str | None = None
-    mapped_to: ArtifactLocation | None = Field(None, alias="mappedTo")
+    mapped_to: ArtifactLocation | None = None
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 # ---------------------------------------------------------------------------
@@ -323,16 +330,14 @@ class SarifRun(BaseModel):
     tool: Tool
     results: list[Result] = []
     invocations: list[Invocation] = []
-    version_control_provenance: list[VersionControlDetails] = Field(
-        [], alias="versionControlProvenance"
-    )
-    automation_details: dict[str, Any] | None = Field(None, alias="automationDetails")
+    version_control_provenance: list[VersionControlDetails] = []
+    automation_details: dict[str, Any] | None = None
     """Optional ``{id, description}`` for fleet-level run identification."""
 
     properties: dict[str, Any] = Field(default_factory=dict)
     """``atlan/profileVersion`` + ``atlan/summary`` totals."""
 
-    model_config = {"populate_by_name": True}
+    model_config = _SARIF_COMMON
 
 
 class SarifReport(BaseModel):
@@ -343,10 +348,13 @@ class SarifReport(BaseModel):
     """
 
     version: str = "2.1.0"
-    schema_: str = Field(
-        "https://raw.githubusercontent.com/oasis-tcs/sarif-spec/master/Schemata/sarif-schema-2.1.0.json",
-        alias="$schema",
-    )
     runs: list[SarifRun] = []
 
-    model_config = {"populate_by_name": True}
+    # version/runs need no camelCase aliases; $schema is injected by the serializer.
+    model_config = ConfigDict(populate_by_name=True)
+
+    @model_serializer(mode="wrap")
+    def _inject_schema(self, handler: Any) -> dict[str, Any]:
+        data: dict[str, Any] = handler(self)
+        data["$schema"] = _SARIF_SCHEMA_URL
+        return data
